requirements.txt
absl-py==1.4.0
alembic==1.11.1
astunparse==1.6.3
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
chex==0.1.7
cloudpickle==1.6.0
cmaes==0.9.1
cmake==3.26.3
colorlog==6.7.0
contourpy==1.0.7
crafter==1.8.1
cycler==0.11.0
decorator==5.1.1
dm-env==1.5
dm-tree==0.1.8
docopt==0.6.2
einops==0.6.1
Farama-Notifications==0.0.4
filelock==3.12.0
flatbuffers==23.5.9
fonttools==4.39.4
gast==0.4.0
gdown==4.7.1
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
gym==0.19.0
gymnasium==0.28.1
gymnasium-notices==0.0.1
h5py==3.8.0
idna==3.4
imageio==2.29.0
-e git+https://github.com/google/jax@ea370435773234f3bfb4bfc9c26af85a13136853#egg=jax
jax-jumpy==1.0.0
jaxlib==0.4.10+cuda11.cudnn86
Jinja2==3.1.2
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lit==16.0.5
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
Minigrid==2.1.0
ml-dtypes==0.1.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvcc-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.6.0.163
nvidia-cudnn-cu12==8.9.1.23
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
opensimplex==0.4.4
opt-einsum==3.3.0
optax==0.1.5
optuna==3.1.1
packaging==23.1
Pillow==9.5.0
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycolab==1.2
pygame==2.1.2
pyglet==2.0.7
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
sympy==1.12
tblib==1.7.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-cpu==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-probability==0.20.1
termcolor==2.3.0
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.0
urllib3==1.26.16
Werkzeug==2.3.4
wrapt==1.14.1
